+++ src/Aggregates.NET/Internal/StreamPosition.cs
@@ -11,10 +11,10 @@ namespace Aggregates.Internal
         //
         // Summary:
         //     The last event in the stream.
-        public const int End = -1;
+        public const long End = -1;
         //
         // Summary:
         //     The first event in a stream
-        public const int Start = 0;
+        public const long Start = 0;
     }
 }
